+++ tests/test_api.py
@@ -56,6 +56,7 @@ def test_basic_1():
             assert results['statistics']['count_ownership_or_control_statement_interest_statement_types'][k] == 1
         else:
             assert results['statistics']['count_ownership_or_control_statement_interest_statement_types'][k] == 0
+    assert results['statistics']['count_replaces_statements_missing'] == 0
 
 
 def test_basic_1_with_birth_year_to_early():
@@ -872,3 +873,51 @@ def test_basic_basic_entity_with_id_with_scheme_name_1():
             assert results['statistics']['count_ownership_or_control_statement_interest_statement_types'][k] == 1
         else:
             assert results['statistics']['count_ownership_or_control_statement_interest_statement_types'][k] == 0
+
+
+def test_replaces_statements():
+
+    cove_temp_folder = tempfile.mkdtemp(prefix='lib-cove-bods-tests-', dir=tempfile.gettempdir())
+    json_filename = os.path.join(os.path.dirname(
+        os.path.realpath(__file__)), 'fixtures', 'api', 'basic_replaces_statements.json'
+    )
+
+    results = bods_json_output(cove_temp_folder, json_filename)
+
+    assert results['validation_errors_count'] == 0
+    assert results['additional_fields_count'] == 0
+    assert results['additional_checks_count'] == 0
+    assert results['file_type'] == 'json'
+    assert results['statistics']['count_entity_statements'] == 2
+    for k in results['statistics']['count_entity_statements_types']:
+        if k == 'registeredEntity':
+            assert results['statistics']['count_entity_statements_types'][k] == 2
+        else:
+            assert results['statistics']['count_entity_statements_types'][k] == 0
+    for k in results['statistics']['count_entity_statements_types_with_any_identifier']:
+        if k == 'registeredEntity':
+            assert results['statistics']['count_entity_statements_types_with_any_identifier'][k] == 2
+        else:
+            assert results['statistics']['count_entity_statements_types_with_any_identifier'][k] == 0
+    for k in results['statistics']['count_entity_statements_types_with_any_identifier_with_id_and_scheme']:
+        if k == 'registeredEntity':
+            assert results['statistics']['count_entity_statements_types_with_any_identifier_with_id_and_scheme'][k] == 2 # noqa
+        else:
+            assert results['statistics']['count_entity_statements_types_with_any_identifier_with_id_and_scheme'][k] == 0 # noqa
+
+    assert results['statistics']['count_person_statements'] == 1
+    for k in results['statistics']['count_person_statements_types']:
+        if k == 'knownPerson':
+            assert results['statistics']['count_person_statements_types'][k] == 1
+        else:
+            assert results['statistics']['count_person_statements_types'][k] == 0
+    assert results['statistics']['count_ownership_or_control_statement'] == 2
+    assert results['statistics']['count_ownership_or_control_statement_interested_party_with_person'] == 2
+    assert results['statistics']['count_ownership_or_control_statement_interested_party_with_entity'] == 0
+    assert results['statistics']['count_ownership_or_control_statement_interested_party_with_unspecified'] == 0
+    for k in results['statistics']['count_ownership_or_control_statement_interest_statement_types']:
+        if k == 'shareholding':
+            assert results['statistics']['count_ownership_or_control_statement_interest_statement_types'][k] == 2
+        else:
+            assert results['statistics']['count_ownership_or_control_statement_interest_statement_types'][k] == 0
+    assert results['statistics']['count_replaces_statements_missing'] == 3